feat: Implement logging to file
This commit introduces the ability to log application output to a file, configurable via a new command-line flag `--log`.

The following changes were made:

- Added a `LogPath` field to the `CLI` struct to allow users to specify the path to the log file via
the command line.
- Implemented the `InitializeLogger` function in the `pkg/log` package, which initializes the global
slog logger to write to either stdout or a specified file.
- Modified the `main` function to call `InitializeLogger` with the `LogPath` value. If the `LogPath`
is empty, logging defaults to stdout.
- Added error handling to `InitializeLogger` to fallback to stdout logging if opening the specified
log file fails, and returns error to main function.