feat: Remove unnecessary print statement in fetch command
This commit removes the `fmt.Println("wait...")` statement from the `Run` method of the `FetchCmd`
struct in `cmd/fetch.go`. This print statement was used for debugging purposes and is no longer
needed, as it provides no valuable information to the user. Removing it cleans up the output and
improves the user experience.