feat: Open article URL in browser from summary view
This commit adds the ability to open the article URL in the default web browser from the summary
view.  A new keybinding "o" has been added to the summary view to trigger this action.  The
webbrowser library is used to open the URL.  An error message is logged if opening the URL fails.