feat: Allow setting speaking rate via config and command line
This commit introduces the ability to configure the speaking rate for text-to-speech functionality
through both a configuration file and command-line arguments.

- The `SpeakingRate` field has been added to the `Config` struct in `config/config.go`, allowing
users to specify a default speaking rate in the configuration file.  A default value of 1.3 is set
if no value is provided.
- The `SpeakingRate` field in `PlayCmd` and `ReadCmd` structs in `cmd/play.go` and `cmd/read.go`
respectively, has been changed to a pointer to a float64 (`*float64`). This allows us to
differentiate between the user not setting the flag and explicitly setting it to the default value.
- The `Run` methods in `cmd/play.go` and `cmd/read.go` now check if the `SpeakingRate`