feat: Add URL method to QueueItem interface and implementations
This commit introduces a new `URL` method to the `QueueItem` interface. This method is implemented
by both `Article` and `playArticle` structs, returning the URL of the article. This change aims to
provide a standardized way to access the URL of an article within the application, improving code
consistency and readability.