feat: Add article list reload functionality
This commit introduces the ability to reload articles in the article list view by pressing the 'r'
key.  This allows users to refresh the article list and fetch the latest updates without navigating
back to the feed list. A debug log is also added to track when the reload function is triggered.